+++ packages/react/src/hooks/index.ts
@@ -43,4 +43,7 @@ export * from "./useQueryString";
 export * from "./useUsdLocked";
 export * from "./useCapInfo";
 export * from "./useTokenBondingKey";
+export * from "./useBondedTokenPrice";
+export * from "./useBondedTokenPrice";
+export * from "./useJupiterPrice";
 export * from "./useManyToOneSwapDriver";