+++ packages/spl-utils/src/accountFetchCache/accountFetchCache.tsx
@@ -194,12 +194,16 @@ export class AccountFetchCache {
         this.commitment
       );
       keys.forEach((key, index) => {
-        this.pendingCallbacks.get(key)!(array[index], null);
+        const callback = this.pendingCallbacks.get(key);
+        callback && callback(array[index], null);
       });
 
       return { keys, array };
     } catch (e: any) {
-      currentBatch.forEach((key) => this.pendingCallbacks.get(key)!(null, e));
+      currentBatch.forEach((key) => {
+        const callback = this.pendingCallbacks.get(key);
+        callback && callback(null, e)
+      });
       throw e;
     }
   }